requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
xgboost
joblib
python-dotenv
transformers
torch
supabase-py
